New functionality : EntityGraph , StoredProcedure Enhancement :  OverrideView
Fixed issue # : DiscriminatorColumn
https://java.net/bugzilla/show_bug.cgi?id=7412
https://java.net/bugzilla/show_bug.cgi?id=7417


git-svn-id: https://svn.java.net/svn/jpamodeler~src/trunk@75 492caa68-1ecd-eb40-82ab-d3e93c0ecf44

+++ JPA Modeler Suite/JPA Modeler/src/org/netbeans/jpa/modeler/core/widget/JavaClassWidget.java
@@ -107,10 +107,10 @@ public abstract class JavaClassWidget extends FlowNodeWidget {
     }
 
     private Image getErrorIcon() {
-        int iconWidth = (int) ((BufferedImage) icon).getWidth() + 3;
-        int iconHeight = (int) ((BufferedImage) icon).getHeight() + 3;
-        int errorIconWidth = (int) ((BufferedImage) errorIcon).getWidth();
-        int errorIconHeight = (int) ((BufferedImage) errorIcon).getHeight();
+        int iconWidth = ((BufferedImage) icon).getWidth() + 3;
+        int iconHeight = ((BufferedImage) icon).getHeight() + 3;
+        int errorIconWidth = ((BufferedImage) errorIcon).getWidth();
+        int errorIconHeight = ((BufferedImage) errorIcon).getHeight();
 
         BufferedImage combined = new BufferedImage(iconWidth, iconHeight, BufferedImage.TYPE_INT_ARGB);
         Graphics g = combined.getGraphics();
